Struct bitmaptrie::BorrowSync

pub struct BorrowSync<'a, T: 'a + Send + Sync> {
    // some fields omitted
}

Borrows a Trie exposing a Sync-type-safe subset of it's methods. No structure modifying methods are included. Each borrow gets it's own path cache. The lifetime of this type is the lifetime of the mutable borrow. This can be used to parallelize structure-immutant changes.

Methods

impl<'a, T: Send + Sync> BorrowSync<'a, T>

fn get(&self, index: usize) -> Option<&'a T>

Retrieve a reference to the value at the given index. Updates the local path cache to point at this index.

fn borrow_sharded(&self, n: usize) -> BorrowShardImmut<'a, T>

Return a type that can be iterated over to produce interior nodes that themselves can be iterated over. This type is immutable.