Struct bitmaptrie::ShardMut

pub struct ShardMut<'a, T: 'a + Send> {
    // some fields omitted
}

A type that references an interior Trie node. For splitting a Trie into sub-nodes, each of which can be passed to a different thread for mutable structural changes.

Methods

impl<'a, T: 'a + Send> ShardMut<'a, T>

fn iter(&self) -> Iter<T>

Return an iterator across this sub tree

fn iter_mut(&mut self) -> IterMut<T>

Return an iterator across mutable values of this sub tree

fn retain_if<F>(&mut self, f: F) where F: FnMut(usize, &mut T) -> bool

Retains only the elements specified by the predicate f.